+++ src/main/java/chess/dao/ConnectionProvider.java
@@ -1,4 +1,4 @@
-package chess.repository;
+package chess.dao;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
