+++ src/main/java/chess/dao/JdbcTemplate.java
@@ -1,4 +1,4 @@
-package chess.repository;
+package chess.dao;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
