+++ src/main/java/chess/dao/RowMapper.java
@@ -1,4 +1,4 @@
-package chess.repository;
+package chess.dao;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
